% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_helpers.R
\name{adjust_raw_scores}
\alias{adjust_raw_scores}
\title{Convert raw scores to adjusted raw scores}
\usage{
adjust_raw_scores(df = NULL, raw, inc = 1)
}
\arguments{
\item{df}{An optional data frame containing a variable for raw scores}

\item{raw}{A list of raw scores, or a quoted column name from the data frame
where raw scores are stored.}

\item{inc}{A value used to increment raw scores in order to calculate adjusted
scores, or an anonymous function}
}
\value{
A list of adjusted raw scores
}
\description{
Convert raw scores to adjusted raw scores
}
\examples{
# Create raw data
df <- data.frame(Id = list(1:20),
RawScore = rep(11:15, 4))
# Increment scores by 2
df$AdjScore <- adjust_raw_scores(df, "RawScore", inc = 2)
}
