% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_comp.R
\name{idmact_comp}
\alias{idmact_comp}
\title{Interpreting Composite Differences in Mean ACT Scores}
\usage{
idmact_comp(
  df = NULL,
  df_map = NULL,
  raw,
  inc,
  map_raw,
  map_scale,
  mcent_subj = function(x) mean(x, na.rm = TRUE),
  mcent_obs = function(x) round(sum(x)/length(x)),
  mcent_comp = function(x) mean(x, na.rm = TRUE)
)
}
\arguments{
\item{df}{A data frame containing raw scores (optional)}

\item{df_map}{A data frame containing the map between raw and scale scores}

\item{raw}{A nested list of raw scores or the list of quoted data frame column
name where raw scores for each subject are stored.}

\item{inc}{A list of numbers. Increment raw scores for each subject by inc to
obtain adjusted scores}

\item{map_raw}{A nested list containing the domain of raw scores for each subject's
raw score to scale score map, or the corresponding list of column names (quoted)
in df or df_map}

\item{map_scale}{A nested list containing the image of scale scores for each
subject's raw score to scale score map, or the corresponding list of column
names (quoted) in df or df_map}

\item{mcent_subj}{A measure of central tendency to summarize subject level
scale scores}

\item{mcent_obs}{An anonymous function to summarize observation level
scale scores (i.e., summary within observation/examinee across subjects)}

\item{mcent_comp}{A anonymous function defining the measure of central tendency
used to summarize composite level scale scores (i.e., summarize exam level scale
score across observations)}
}
\value{
A nested list
}
\description{
Interpreting Composite Differences in Mean ACT Scores
}
\examples{
raw = list(list(1, 2, 3, 4, 5), list(1, 1, 1, 1, 1))
inc = list(1 , 1)
map_raw = list(list(1, 2, 3, 4, 5))
map_scale = list(list(20, 21, 22, 23, 24))
comp_mean <- idmact_comp(raw = raw,
                         inc = inc,
                         map_raw = map_raw,
                         map_scale = map_scale)
}
