% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_subj.R
\name{idmact_subj}
\alias{idmact_subj}
\title{Interpreting Differences in Mean ACT Scores at the Subject Level}
\usage{
idmact_subj(
  df = NULL,
  df_map = NULL,
  raw,
  inc = 1,
  map_raw,
  map_scale,
  mcent_subj = function(x) mean(x, na.rm = TRUE)
)
}
\arguments{
\item{df}{An optional data frame containing a variable for raw scores}

\item{df_map}{A data frame that maps raw scores to their corresponding scale
scores}

\item{raw}{A list of raw scores, or a quoted column name from the data frame
where raw scores are stored.}

\item{inc}{A value used to increment raw scores in order to calculate adjusted
scores}

\item{map_raw}{A list containing the domain of raw scores for the raw-to-scale
score mapping, or a quoted column name from either df or df_map that represents
this domain}

\item{map_scale}{A list containing the range of scale scores for the
raw-to-scale score mapping, or a quoted column name from either df or df_map
that represents this range}

\item{mcent_subj}{An anonymous function that defines the measure of central
tendency to be used}
}
\value{
A nested list containing betas, a summary of scale scores as a list
(adjusted and unadjusted), a list of adjusted and unadjusted scale scores,
and a list of adjusted and unadjusted raw scores
}
\description{
Interpreting Differences in Mean ACT Scores at the Subject Level
}
\examples{
raw = list(1, 2, 3, 4, 5)
map_raw = list(1, 2, 3, 4, 5)
map_scale = list(20, 21, 22, 23, 24)
idmact_subj(raw = raw,
            map_raw = map_raw,
            map_scale = map_scale)
}
