% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_helpers.R
\name{map_elongate}
\alias{map_elongate}
\title{Elongate Map}
\usage{
map_elongate(map_raw, map_scale)
}
\arguments{
\item{map_raw}{A list containing the domain of raw scores for the raw-to-scale
score mapping}

\item{map_scale}{A list containing the domain of raw scores for the raw-to-scale
score mapping}
}
\value{
A nested list consisting of a list of raw scores and a list of scale
scores.
}
\description{
Elongate Map
}
\examples{
map_elongate(map_raw = list(1:5, 6:10),
map_scale = list(20, 21))
}
