% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_helpers.R
\name{map_elongate_df}
\alias{map_elongate_df}
\title{Elongate df_map}
\usage{
map_elongate_df(df_map, map_raw, map_scale)
}
\arguments{
\item{df_map}{A data frame that maps raw scores to their corresponding scale
scores}

\item{map_raw}{A quoted column name from df_map containing the domain of raw
scores for the raw-to-scale score mapping}

\item{map_scale}{A quoted column name from df_map containing the range of
scale scores for the raw-to-scale score mapping}
}
\value{
A data frame
}
\description{
Elongate df_map
}
\examples{
df_map <- data.frame(
 raw = c("1-5", "6-10"),
 scale = c(10, 11))
# Elongate the data frame
elongated_df <- map_elongate_df(df_map, "raw", "scale")
}
