% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_helpers.R
\name{map_scores}
\alias{map_scores}
\title{Convert Raw Scores to Scale Scores}
\usage{
map_scores(
  df = NULL,
  df_map = NULL,
  conv,
  map_raw,
  map_scale,
  na.rm.max = TRUE
)
}
\arguments{
\item{df}{A data frame containing raw scores (optional)}

\item{df_map}{A data frame containing the map between raw and scale scores}

\item{conv}{A list of raw scores or the column name (quoted) of the scores in df}

\item{map_raw}{A list containing the domain of raw scores for the raw score to
scale score map, or the corresponding column name (quoted) in df or df_map}

\item{map_scale}{A list containing the image of scale scores for the raw score
to scale score map, or the corresponding column name (quoted) in df or df_map}

\item{na.rm.max}{Pass na.rm argument to max in order to compute maximum raw
and scale values of the map}
}
\value{
A list (if conv is a list) or a data frame (if conv is a column name)
}
\description{
Convert Raw Scores to Scale Scores
}
\examples{
(bij_scale <- map_scores(conv = list(1, 2, 3, 4, 5),
                        map_raw = list(1, 2, 3, 4, 5),
                        map_scale = list(20, 21, 22, 23, 24)))
}
