% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_helpers.R
\name{adjust_raw_scores}
\alias{adjust_raw_scores}
\title{idmact Helper Functions}
\usage{
adjust_raw_scores(df = NULL, raw, inc = 1)
}
\arguments{
\item{df}{A data frame}

\item{raw}{A list of raw scores or the quoted data frame column name where
raw scores are stored.}

\item{inc}{Increment raw scores by inc to obtain adjusted scores}
}
\value{
A list
}
\description{
idmact Helper Functions
}
\examples{
# Create raw data
df <- data.frame(Id = list(1:20),
RawScore = rep(11:15, 4))
# Increment scores by 2
df$AdjScore <- adjust_raw_scores(df, "RawScore", inc = 2)
}
