% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_comp.R
\name{idmact_comp}
\alias{idmact_comp}
\title{Interpreting Differences in Mean ACT Scores at the Composite Level}
\usage{
idmact_comp(
  df = NULL,
  df_map = NULL,
  raw,
  inc,
  map_raw,
  map_scale,
  mcent_subj = function(x) mean(x, na.rm = TRUE),
  mcent_obs = function(x) round(sum(x)/length(x)),
  mcent_comp = function(x) mean(x, na.rm = TRUE)
)
}
\arguments{
\item{df}{An optional data frame containing a variable for raw scores}

\item{df_map}{A data frame that maps raw scores to their corresponding scale
scores}

\item{raw}{A list containing either a list of raw scores for each subject, or
quoted column names from the data frame where raw scores for each subject are
stored}

\item{inc}{A value used to increment raw scores in order to calculate adjusted
scores, or an anonymous function}

\item{map_raw}{A nested list where each sublist contains the domain of raw
scores for a subject's raw-to-scale score mapping, or quoted column names from
either df or df_map representing the subject area domains}

\item{map_scale}{A nested list where each sublist contains the range of scale
scores for a subject's raw-to-scale score mapping, or quoted column names from
either df or df_map representing the subject area ranges}

\item{mcent_subj}{An anonymous function specifying the measure of central
tendency used to summarize scale scores at the subject level}

\item{mcent_obs}{An anonymous function used to summarize scale scores at the
observation level (i.e., summary within each observation/examinee across
subjects)}

\item{mcent_comp}{An anonymous function defining the measure of central
tendency used to summarize composite level scale scores (i.e., summarizing
exam level scale scores across observations)}
}
\value{
A nested list containing both composite and subject level results.
Composite results include deltac, a list of summarized composite scale scores
(adjusted and unadjusted), and composite level scale scores (adjusted and
unadjusted). Subject level results consist of the outcomes obtained from
idmact_subj for each subject.
}
\description{
Use the idmact_comp() function to interpret differences in composite level
scores.  The algorithm is implemented as: 1) Increment the raw score for one
or more subjects for each student to obtain adjusted raw scores; 2) Map
adjusted raw scores to adjusted scale scores using the form's raw score to
scale score map (note: perfect raw scores are always converted to the maximum
allowable scale score despite the adjustment in step one); 3) Sum the adjusted
scale scores for each subject area, divide this sum by the number of subject
areas, and round to the nearest integer in order to obtain each
observation/examinee's adjusted composite scale score; 4) Calculate the
adjusted mean composite scale score across all observations (m_adj);
5) Calculate the unadjusted mean composite scale score across all observations
(m_unadj); 6) Compute the difference between the adjusted and unadjusted mean
composite scale scores to obtain delta composite: deltac = m_adj - m_unadj
}
\details{
The default idmact_comp() parameter values for inc, mcent_subj, mcent_obs, and
mcent_comp provide an implementation of the algorithm aligned with the main
method presented in Schiel (1998)
\url{https://www.act.org/content/dam/act/unsecured/documents/ACT_RR98-01.pdf};
however, these parameters can also take arbitrary anonymous functions for users
wishing to use modified implementations of the algorithm.
}
\examples{
raw = list(list(1, 2, 3, 4, 5), list(1, 1, 1, 1, 1))
inc = list(1 , 1)
map_raw = list(list(1, 2, 3, 4, 5))
map_scale = list(list(20, 21, 22, 23, 24))
comp_mean <- idmact_comp(raw = raw,
                         inc = inc,
                         map_raw = map_raw,
                         map_scale = map_scale)
}
