% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_comp.R
\name{idmact_comp}
\alias{idmact_comp}
\title{Interpreting Differences in Mean ACT Scores at the Composite Level}
\usage{
idmact_comp(
  df = NULL,
  df_map = NULL,
  raw,
  inc,
  map_raw,
  map_scale,
  mcent_subj = function(x) mean(x, na.rm = TRUE),
  mcent_obs = function(x) round(sum(x)/length(x)),
  mcent_comp = function(x) mean(x, na.rm = TRUE)
)
}
\arguments{
\item{df}{An optional data frame containing a variable for raw scores}

\item{df_map}{A data frame that maps raw scores to their corresponding scale
scores}

\item{raw}{A list containing either a list of raw scores for each subject, or
quoted column names from the data frame where raw scores for each subject are
stored}

\item{inc}{A list of values used to increment raw scores for each subject, in
order to calculate adjusted scores}

\item{map_raw}{A nested list where each sublist contains the domain of raw
scores for a subject's raw-to-scale score mapping, or quoted column names from
either df or df_map representing the subject area domains}

\item{map_scale}{A nested list where each sublist contains the range of scale
scores for a subject's raw-to-scale score mapping, or quoted column names from
either df or df_map representing the subject area ranges}

\item{mcent_subj}{An anonymous function specifying the measure of central
tendency used to summarize scale scores at the subject level}

\item{mcent_obs}{An anonymous function used to summarize scale scores at the
observation level (i.e., summary within each observation/examinee across
subjects)}

\item{mcent_comp}{An anonymous function defining the measure of central
tendency used to summarize composite level scale scores (i.e., summarizing
exam level scale scores across observations)}
}
\value{
A nested list containing both composite and subject level results.
Composite results include betac, a list of summarized composite scale scores
(adjusted and unadjusted), and composite level scale scores (adjusted and
unadjusted). Subject level results consist of the outcomes obtained from
idmact_subj for each subject.
}
\description{
Interpreting Differences in Mean ACT Scores at the Composite Level
}
\examples{
raw = list(list(1, 2, 3, 4, 5), list(1, 1, 1, 1, 1))
inc = list(1 , 1)
map_raw = list(list(1, 2, 3, 4, 5))
map_scale = list(list(20, 21, 22, 23, 24))
comp_mean <- idmact_comp(raw = raw,
                         inc = inc,
                         map_raw = map_raw,
                         map_scale = map_scale)
}
