% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_subj.R
\name{idmact_subj}
\alias{idmact_subj}
\title{Interpreting Subject Level Differences in Mean ACT Scores}
\usage{
idmact_subj(
  df = NULL,
  df_map = NULL,
  raw,
  inc = 1,
  map_raw,
  map_scale,
  mcent = "mean",
  na.rm.mcent = TRUE
)
}
\arguments{
\item{df}{A data frame containing raw scores (optional)}

\item{df_map}{A data frame containing the map between raw and scale scores}

\item{raw}{A list of raw scores or the quoted data frame column name where
raw scores are stored.}

\item{inc}{Increment raw scores by inc to obtain adjusted scores}

\item{map_raw}{A list containing the domain of raw scores for the raw score to
scale score map, or the corresponding column name (quoted) in df or df_map}

\item{map_scale}{A list containing the image of scale scores for the raw score
to scale score map, or the corresponding column name (quoted) in df or df_map}

\item{mcent}{A measure of central tendency}

\item{na.rm.mcent}{Pass na.rm argument to function selected with mcent in order
to summarize scale score}
}
\value{
A list containing betas, scale score summary (adjusted and unadjusted),
scale scores (adjusted and unadjusted), and raw scores (adjusted nd unadjusted)
}
\description{
Interpreting Subject Level Differences in Mean ACT Scores
}
\examples{
raw = list(1, 2, 3, 4, 5)
map_raw = list(1, 2, 3, 4, 5)
map_scale = list(20, 21, 22, 23, 24)
idmact_subj(raw = raw,
            map_raw = map_raw,
            map_scale = map_scale)
}
