% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmact_helpers.R
\name{map_scores}
\alias{map_scores}
\title{Convert Raw Scores to Scale Scores}
\usage{
map_scores(
  df = NULL,
  df_map = NULL,
  conv,
  map_raw,
  map_scale,
  na.rm.max = TRUE
)
}
\arguments{
\item{df}{An optional data frame containing a variable for raw scores}

\item{df_map}{A data frame that maps raw scores to their corresponding scale
scores}

\item{conv}{A list of raw scores, or a quoted column name from the data frame
where raw scores are stored.}

\item{map_raw}{A list containing the domain of raw scores for the raw-to-scale
score mapping, or a quoted column name from either df or df_map that represents
this domain}

\item{map_scale}{A list containing the range of scale scores for the
raw-to-scale score mapping, or a quoted column name from either df or df_map
that represents this range}

\item{na.rm.max}{Pass the na.rm argument to the max function for computing the
maximum raw and scale values in the mapping, taking into account the handling
of missing values}
}
\value{
A list if 'conv' is a list, or a data frame if 'conv' is a column name
}
\description{
Convert Raw Scores to Scale Scores
}
\examples{
(bij_scale <- map_scores(conv = list(1, 2, 3, 4, 5),
                        map_raw = list(1, 2, 3, 4, 5),
                        map_scale = list(20, 21, 22, 23, 24)))
}
